refactor: split files into ui and internal packages
