+++ src/main/scala/it/scalalearn/goosegame/Main.scala
@@ -1,8 +1,8 @@
 package it.scalalearn.goosegame
 
-import it.scalalearn.goosegame.cli.CLIStrings.START_MSG
-import it.scalalearn.goosegame.cli.CommandLineInterface.cli
-import it.scalalearn.goosegame.gamestate.GameState
+import it.scalalearn.goosegame.internal.gamestate.GameState
+import it.scalalearn.goosegame.ui.cli.CLIStrings.START_MSG
+import it.scalalearn.goosegame.ui.cli.CommandLineInterface.cli
 
 object Main {
   def main(args: Array[String]): Unit = {
