+++ src/main/scala/it/scalalearn/goosegame/internal/movelogic/Move.scala
@@ -1,3 +1,3 @@
-package it.scalalearn.goosegame.movelogic
+package it.scalalearn.goosegame.internal.movelogic
 
 case class Move(name: String, previousSquare: Int, startSquare: Int, dice: List[Int])
