+++ src/main/scala/it/scalalearn/goosegame/internal/movelogic/MoveType.scala
@@ -1,4 +1,4 @@
-package it.scalalearn.goosegame.movelogic
+package it.scalalearn.goosegame.internal.movelogic
 
 enum MoveType {
   case BOUNCE, BRIDGE, GOOSE_START, GOOSE_CONTINUE, LAST, NORMAL
