+++ src/main/scala/it/scalalearn/goosegame/ui/readout/ReadoutBuilder.scala
@@ -1,10 +1,11 @@
-package it.scalalearn.goosegame.readout
+package it.scalalearn.goosegame.ui.readout
 
-import it.scalalearn.goosegame.gamestate.GameState
-import it.scalalearn.goosegame.movelogic.MoveType
-import it.scalalearn.goosegame.movelogic.MoveType.{BOUNCE, BRIDGE, GOOSE_CONTINUE, GOOSE_START, LAST, NORMAL}
-import it.scalalearn.goosegame.readout.ReadoutData
-import it.scalalearn.goosegame.readout.ReadoutMessages.{LIST_PLAYERS_MSG, MID_ROLL_BOUNCE_MSG, MID_ROLL_BRIDGE_MSG, MID_ROLL_GOOSE_CONTINUE_MSG, MID_ROLL_GOOSE_START_MSG, MID_ROLL_PRANK_MSG, START_ROLL_MSG, WIN_MSG}
+import it.scalalearn.goosegame.internal.gamestate.GameState
+import it.scalalearn.goosegame.internal.movelogic.MoveType
+import it.scalalearn.goosegame.internal.movelogic.MoveType.{BOUNCE, BRIDGE, GOOSE_CONTINUE, GOOSE_START, LAST, NORMAL}
+import it.scalalearn.goosegame.ui.readout.ReadoutData
+import it.scalalearn.goosegame.ui.readout.ReadoutMessages.{LIST_PLAYERS_MSG, MID_ROLL_BOUNCE_MSG, MID_ROLL_BRIDGE_MSG,
+  MID_ROLL_GOOSE_CONTINUE_MSG, MID_ROLL_GOOSE_START_MSG, MID_ROLL_PRANK_MSG, START_ROLL_MSG, WIN_MSG}
 
 object ReadoutBuilder {
   def appendMessage(readoutData: ReadoutData, newMessage: String): ReadoutData =
